requirements.txt
# Web framework
Flask==2.3.3
gunicorn==21.2.0


# CORS support
Flask-Cors==6.0.1

# Machine learning 
torch==2.2.0+cpu
torchvision==0.17.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu


# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Requests for API calls
requests==2.31.0

# For environment variable management
python-dotenv==1.0.1

# Optional: for async tasks / background jobs
celery==5.3.1
redis==5.3.0

# Optional: JSON handling
ujson==5.8.0
